Move artifacts to contract config folder
