+++ src/controllers/gas_station/post.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { hex2a, parseLogs, findEvent } from '@/util/events';
 import { TransactionService } from '@/services/TransactionService';
-import { Artifacts } from '@/util/artifacts';
+import { Artifacts } from '@/config/contracts/artifacts';
 import { eventIndexer } from '@/util/indexer';
 import { body } from 'express-validator';
 import { InternalServerError } from '@/util/errors';
