+++ src/util/upgrades.ts
@@ -2,7 +2,7 @@ import { getSelectors, ADDRESS_ZERO } from './network';
 import { TransactionService } from '@/services/TransactionService';
 import { poolFacetAdresses, poolFacetContracts } from '@/config/contracts';
 import { pick, uniq } from '.';
-import { Artifacts } from './artifacts';
+import { Artifacts } from '../config/contracts/artifacts';
 import { AssetPoolDocument, AssetPoolType } from '@/models/AssetPool';
 
 export enum FacetCutAction {
